% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scRNA.R
\name{batch_correct_pipeline}
\alias{batch_correct_pipeline}
\title{Batch correct method}
\usage{
batch_correct_pipeline(
  object_list,
  method,
  normalization.method = "LogNormalize",
  scale.factor = 10000,
  selection.method = "vst",
  nfeatures = 2000,
  mean.cutoff = c(0.1, 8),
  dispersion.cutoff = c(1, Inf),
  reference = NULL,
  dims = 1:30,
  k.anchor = 5,
  k.filter = 200,
  k.score = 30,
  k.weight = 100,
  features = NULL,
  features_scale = NULL,
  vars.to.regress = NULL,
  npcs = 50,
  k.param = 50,
  resolution = 0.8,
  n.neighbors = 30,
  min.dist = 0.3
)
}
\arguments{
\item{object_list}{A list contains seurat object.}

\item{method}{Select a method to correct the batch effect among list, eg. seurat, harmony, and liger. Default: seurat.}

\item{normalization.method}{Method for normalization. Default: LogNormalize.}

\item{scale.factor}{Sets the scale factor for cell-level normalization. Default: 10000.}

\item{selection.method}{How to choose top variable features. Default: vst.}

\item{nfeatures}{Number of features to select as top variable features among list, Default:2000.}

\item{mean.cutoff}{A two-length numeric vector with low- and high-cutoffs for feature means. Default: c(0.1, 8).}

\item{dispersion.cutoff}{A two-length numeric vector with low- and high-cutoffs for feature dispersions. Default: c(1, Inf).}

\item{reference}{A vector specifying the object/s to be used as a reference during integration. If NULL (default), all pairwise anchors are found (no reference/s). If not NULL, the corresponding objects in object.list will be used as references. When using a set of specified references, anchors are first found between each query and each reference. The references are then integrated through pairwise integration. Each query is then mapped to the integrated reference.}

\item{dims}{Dimensions of reduction to use as input. Default: 1:10.}

\item{k.anchor}{How many neighbors (k) to use when picking anchors, Default:5.}

\item{k.filter}{How many neighbors (k) to use when filtering anchors, Default:200.}

\item{k.score}{How many neighbors (k) to use when scoring anchors, Default:30.}

\item{k.weight}{Number of neighbors to consider when weighting anchors, Default:100.}

\item{features}{Vector of features names to scale/center. Default is variable features. Default: NULL.}

\item{features_scale}{Vector of features names to scale/center. Default is variable features. Default: NULL.}

\item{vars.to.regress}{Variables to regress out (previously latent.vars in RegressOut). For example, nUMI, or percent.mito. Default: NULL.}

\item{npcs}{Total Number of PCs to compute and store (50 by default).}

\item{k.param}{Defines k for the k-nearest neighbor algorithm. Default: 50.}

\item{resolution}{Value of the resolution parameter, use a value above (below) 1.0 if you want to obtain a larger (smaller) number of communities. Default: 0.8.}

\item{n.neighbors}{This determines the number of neighboring points used in local approximations of manifold structure. Larger values will result in more global structure being preserved at the loss of detailed local structure. In general this parameter should often be in the range 5 to 50.}

\item{min.dist}{This controls how tightly the embedding is allowed compress points together. Larger values ensure embedded points are moreevenly distributed, while smaller values allow the algorithm to optimise more accurately with regard to local structure. Sensible values are in the range 0.001 to 0.5.}
}
\value{
The batch corrected Seurat object.
}
\description{
Batch correct method
}
