% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scRNA.R
\name{reference_mapping}
\alias{reference_mapping}
\title{Automatic celltype annotation through reference mapping}
\usage{
reference_mapping(object, reference)
}
\arguments{
\item{object}{A seurat object needs to be annotated.}

\item{reference}{A reference seurat object containing celltype info.}
}
\value{
The annotated Seurat object.
}
\description{
Automatic celltype annotation through reference mapping
}
